Move xHCI controller init into Root__New
